+++ app/src/main/java/com/agile/app/repository/ApplicationRepository.java
@@ -3,5 +3,5 @@ package com.agile.app.repository;
 import com.agile.app.entity.Application;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApplicationRepository extends JpaRepository<Application, Long> {
+public interface ApplicationRepository extends JpaRepository<Application, Integer> {
 }
